# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "커피 키오스크 API",
    "description": "키오스크·백오피스·마이페이지 API. 인증 필요 엔드포인트는 Authorize에 JWT를 입력하세요.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "로컬 개발" }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "회원가입",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "username", "email", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "username": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "가입 성공", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterResponse" } } } },
          "400": { "description": "필수값 누락 또는 비밀번호 6자 미만" },
          "409": { "description": "username 또는 email 중복" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "로그인 (Access 30분 + Refresh 7일 발급)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "성공", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } } },
          "400": { "description": "username, password 필수" },
          "401": { "description": "invalid username or password" }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh 토큰으로 Access 토큰 재발급 (Refresh 로테이션)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": { "refreshToken": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "성공", "content": { "application/json": { "schema": { "type": "object", "properties": { "accessToken": { "type": "string" }, "refreshToken": { "type": "string" }, "expiresIn": { "type": "number" } } } } } },
          "400": { "description": "refreshToken required" },
          "401": { "description": "invalid or expired refresh token" }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh 토큰 폐기",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "refreshToken": { "type": "string" } } }
            }
          }
        },
        "responses": { "204": { "description": "성공" } }
      }
    },
    "/api/auth/find-id": {
      "post": {
        "tags": ["Auth"],
        "summary": "아이디 찾기 (성함+이메일)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email"],
                "properties": { "name": { "type": "string" }, "email": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "성공", "content": { "application/json": { "schema": { "type": "object", "properties": { "username": { "type": "string" } } } } } },
          "400": { "description": "name, email required" },
          "404": { "description": "no user found" }
        }
      }
    },
    "/api/auth/find-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "비밀번호 찾기 (임시 비밀번호 발급, 성공 여부만 반환)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "username", "email"],
                "properties": { "name": { "type": "string" }, "username": { "type": "string" }, "email": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "성공", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "name, username, email required" },
          "404": { "description": "no user found" }
        }
      }
    },
...
    "/api/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "활성 카테고리 목록",
        "responses": {
          "200": { "description": "성공", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Category" } } } } }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "카테고리 등록",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "isActive": { "type": "boolean" },
                  "sortOrder": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "등록 성공" },
          "400": { "description": "name required" }
        }
      }
    },
    "/api/categories/{id}": {
      "patch": {
        "tags": ["Categories"],
        "summary": "카테고리 수정",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "isActive": { "type": "boolean" },
                  "sortOrder": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "성공" } }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "카테고리 삭제",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "204": { "description": "삭제 완료" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "로그인 응답의 token 값"
      }
    },
    "schemas": {
...
      "RegisterResponse": {
        "type": "object",
        "properties": { "user": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "username": { "type": "string" }, "email": { "type": "string" }, "point": { "type": "number" }, "mileage": { "type": "number" } } } }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "expiresIn": { "type": "number", "description": "Access 만료까지 초 단위 (1800=30분)" },
          "user": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "name": { "type": "string" },
              "username": { "type": "string" },
              "email": { "type": "string" },
              "role": { "type": "string", "enum": ["USER", "ADMIN"] },
              "point": { "type": "number" },
              "mileage": { "type": "number" }
            }
          }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "sortOrder": { "type": "number" }
        }
      }
    }
  }
}